(nodemapper) refactor builder query to backend
