+++ nodemapper/src/redux/middleware/builder.ts
@@ -5,10 +5,10 @@ import { DefaultLinkModel } from "NodeMap";
 import { DefaultNodeModel } from "NodeMap";
 
 import { builderRedraw } from "redux/actions";
-import { builderSubmitQuery } from "redux/actions";
 import { builderCompileToJson } from "redux/actions";
 import { builderUpdateNodeInfo } from "redux/actions";
 import { builderUpdateStatusText } from "redux/actions";
+import { builderUpdateModulesList } from "redux/actions";
 
 // TODO: Replace with webpack proxy (problems getting this to work)
 const API_ENDPOINT = "http://127.0.0.1:5000/api";
@@ -35,10 +35,13 @@ export function builderMiddleware({ getState, dispatch }) {
           NodeDeselected(dispatch);
           break;
         case "builder/get-remote-modules":
-          GetRemoteModules(dispatch, dispatch, getState().builder.repo);
+          GetRemoteModules(dispatch, getState().builder.repo);
           break;
         case "builder/update-modules-list":
           UpdateModulesList(dispatch);
+          break;
+        case "builder/import-module":
+          ImportModule();
           break;
         default:
           break;
@@ -150,7 +153,6 @@ function NodeDeselected(dispatch: TNodeDeselectedDispatch) {
 }
 
 function GetRemoteModules(
-  dispatchSubmitQuery: TPayloadRecord,
   dispatchUpdateStatusText: TPayloadString,
   repo: string
 ) {
@@ -168,12 +170,16 @@ function GetRemoteModules(
       }),
     },
   };
-  dispatchSubmitQuery(builderSubmitQuery(query));
+  SubmitQuery(query, dispatchUpdateStatusText);
 }
 
 function UpdateModulesList(dispatch: TPayloadString) {
   // Update list of modules - done in reducer
   dispatch(builderUpdateStatusText(""));
+}
+
+function ImportModule() {
+  // Query user for config file and import module
 }
 
 ///////////////////////////////////////////////////////////////////////////////
@@ -277,3 +283,52 @@ async function postRequestCheckNodeDependencies(
       console.error("Error during query: ", error);
     });
 }
+
+const SubmitQuery = (query: Record<string, any>, dispatch) => {
+  // POST request handler
+  async function postRequest() {
+    const postRequestOptions = {
+      method: "POST",
+      headers: { "Content-Type": "application/json;charset=UTF-8" },
+      body: JSON.stringify(query),
+    };
+    console.info("Sending query: ", query);
+    fetch(API_ENDPOINT + "/post", postRequestOptions)
+      .then((response) => {
+        if (response.ok) {
+          return response.json();
+        }
+        dispatch(builderUpdateStatusText("Error: " + response.statusText));
+        throw response;
+      })
+      .then((data) => {
+        if (data !== null) {
+          processResponse(data);
+        }
+        console.info("Got response: ", data);
+      })
+      .catch((error) => {
+        console.error("Error during query: ", error);
+      });
+  }
+
+  function processResponse(content: JSON) {
+    console.log("Process response: ", content);
+    dispatch(builderUpdateStatusText(""));
+    switch (content["query"]) {
+      case "builder/get-remote-modules": {
+        dispatch(builderUpdateModulesList(content["body"]));
+        break;
+      }
+      default:
+        console.error(
+          "Error interpreting server response (query: ",
+          content["query"],
+          ")"
+        );
+    }
+  }
+
+  // Received query request
+  if (JSON.stringify(query) !== JSON.stringify({})) postRequest();
+};
